+++ programs/dynamic-bonding-curve/src/instructions/operator/mod.rs
@@ -2,3 +2,5 @@ pub mod ix_claim_protocol_fee;
 pub use ix_claim_protocol_fee::*;
 pub mod ix_claim_protocol_pool_creation_fee;
 pub use ix_claim_protocol_pool_creation_fee::*;
+pub mod ix_zap_protocol_fee;
+pub use ix_zap_protocol_fee::*;