+++ tests/instructions/adminInstructions.ts
@@ -14,65 +14,43 @@ import {
 } from "../utils";
 import {
   deriveClaimFeeOperatorAddress,
+  deriveOperatorAddress,
   derivePoolAuthority,
 } from "../utils/accounts";
 import { VirtualCurveProgram } from "../utils/types";
+import BN from "bn.js";
 
-export type CreateClaimProtocolFeeOperatorParams = {
-  admin: Keypair;
-  operator: PublicKey;
-};
+export enum OperatorPermission {
+  ClaimProtocolFee,
+  ZapProtocolFee,
+}
 
-export async function createClaimProtocolFeeOperator(
+export function encodePermissions(permissions: OperatorPermission[]): BN {
+  return permissions.reduce((acc, perm) => {
+    return acc.or(new BN(1).shln(perm));
+  }, new BN(0));
+}
+
+export async function createOperatorAccount(
   svm: LiteSVM,
   program: VirtualCurveProgram,
-  params: CreateClaimProtocolFeeOperatorParams
-): Promise<PublicKey> {
-  const { operator, admin } = params;
-  const claimFeeOperator = deriveClaimFeeOperatorAddress(operator);
+  params: {
+    admin: Keypair;
+    whitelistedAddress: PublicKey;
+    permissions: OperatorPermission[];
+  }
+) {
+  const { admin, whitelistedAddress, permissions } = params;
+
   const transaction = await program.methods
-    .createClaimProtocolFeeOperator()
+    .createOperatorAccount(encodePermissions(permissions))
     .accountsPartial({
-      claimFeeOperator,
-      operator,
       signer: admin.publicKey,
+      operator: deriveOperatorAddress(whitelistedAddress),
+      whitelistedAddress,
       payer: admin.publicKey,
     })
     .transaction();
-
-  sendTransactionMaybeThrow(svm, transaction, [admin]);
-
-  const claimFeeOperatorState = getClaimFeeOperator(
-    svm,
-    program,
-    claimFeeOperator
-  );
-  expect(claimFeeOperatorState.operator.toString()).eq(operator.toString());
-
-  return claimFeeOperator;
-}
-
-export async function closeClaimProtocolFeeOperator(
-  svm: LiteSVM,
-  program: VirtualCurveProgram,
-  admin: Keypair,
-  claimProtocolFeeOperator: PublicKey
-): Promise<any> {
-  const transaction = await program.methods
-    .closeClaimProtocolFeeOperator()
-    .accounts({
-      claimFeeOperator: claimProtocolFeeOperator,
-      rentReceiver: admin.publicKey,
-      signer: admin.publicKey,
-    })
-    .transaction();
-
-  const claimFeeOperatorState = getClaimFeeOperator(
-    svm,
-    program,
-    claimProtocolFeeOperator
-  );
-  expect(claimFeeOperatorState).to.be.null;
 
   sendTransactionMaybeThrow(svm, transaction, [admin]);
 }
@@ -81,7 +59,6 @@ export type ClaimLegacyPoolCreationFeeParams = {
   operator: Keypair;
   pool: PublicKey;
 };
-
 
 export type ClaimProtocolPoolCreationFeeParams = {
   operator: Keypair;
@@ -105,7 +82,7 @@ export async function claimProtocolPoolCreationFee(
       config: poolState.config,
       treasury: TREASURY,
       signer: operator.publicKey,
-      claimFeeOperator,
+      operator: deriveOperatorAddress(operator.publicKey),
     })
     // Trick to bypass bankrun transaction has been processed if we wish to execute same tx again
     .remainingAccounts([
@@ -133,8 +110,7 @@ export async function claimProtocolFee(
   const poolState = getVirtualPool(svm, program, pool);
   const configState = getConfig(svm, program, poolState.config);
   const poolAuthority = derivePoolAuthority();
-  const claimFeeOperator = deriveClaimFeeOperatorAddress(operator.publicKey);
-  const quoteMintInfo = getTokenAccount(svm, poolState.quoteVault);
+  const quoteMintInfo = getTokenAccount(svm, poolState.quoteVault)!;
 
   const tokenBaseProgram =
     configState.tokenType == 0 ? TOKEN_PROGRAM_ID : TOKEN_2022_PROGRAM_ID;
@@ -177,7 +153,7 @@ export async function claimProtocolFee(
       quoteMint: quoteMintInfo.mint,
       tokenBaseAccount,
       tokenQuoteAccount,
-      claimFeeOperator,
+      operator: deriveOperatorAddress(operator.publicKey),
       signer: operator.publicKey,
       tokenBaseProgram,
       tokenQuoteProgram,
@@ -186,6 +162,4 @@ export async function claimProtocolFee(
     .transaction();
 
   sendTransactionMaybeThrow(svm, transaction, [operator]);
-
 }
-